Am facut CRUD pentru Project
+ am mai incercat sa modific CRUD pentru Team mic fail :(

+++ App.NET/Controllers/ProjectsController.cs
@@ -15,16 +15,23 @@ namespace App.NET.Controllers
 
         public IActionResult Index(int team_id)
         {
-            var projects = db.Projects.Include("ApplicationUser").Include("Team")
-                .Where(project => project.User_id == User.Identity.Name && project.Team_id == team_id);
-            return View();
+            var projects = db.Projects
+                        .Include(p => p.Team)
+                        .Where(project => project.User_id == User.Identity.Name && project.Team_id == team_id)
+                        .ToList();
+
+
+            return View(projects);
         }
 
         public IActionResult Show(int id)
         {
             Project project = db.Projects.Find(id);
 
-            ICollection<Task_table> tasks = db.Tasks.Include("Project").Where(task => task.Project_id == id).ToList();
+            ICollection<Task_table> tasks = db.Tasks
+                .Include(t => t.Project)
+                .Where(task => task.Project_id == id)
+                .ToList();
 
             ViewBag.Tasks = tasks;
             return View(project);
@@ -32,11 +39,11 @@ namespace App.NET.Controllers
 
         public IActionResult New()
         {
-            return View();
+            Project newProject = new Project();
+            return View(newProject);
         }
 
         [HttpPost]
-
         public IActionResult New(Project project)
         {
             if (ModelState.IsValid)
@@ -58,38 +65,45 @@ namespace App.NET.Controllers
             return View(project);
         }
 
-        [HttpPut]
-        public IActionResult Edit (int id, Project requestProject)
+        [HttpPost]
+        public IActionResult Edit(int id, Project requestProject)
         {
             Project project = db.Projects.Find(id);
-           
-           
-                if (ModelState.IsValid)
-                {
-                    project.title = requestProject.title;
-                    project.description = requestProject.description;
-                    db.SaveChanges();
-                    TempData["message"] = "Proiectul a fost modificat!";
 
-                    return RedirectToAction("Index");
-                }
-                else
-                {
-                    return View(requestProject);
-                }
-            
-         
+            if (project == null)
+            {
+                return NotFound();
+            }
+
+            if (ModelState.IsValid)
+            {
+                project.Title_project = requestProject.Title_project;
+                project.Description = requestProject.Description;
+                db.SaveChanges();
+                TempData["message"] = "Proiectul a fost modificat!";
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return View(requestProject);
+            }
         }
 
         [HttpPost]
         public IActionResult Delete(int id)
         {
             Project project = db.Projects.Find(id);
+
+            if (project == null)
+            {
+                return NotFound();
+            }
+
             db.Projects.Remove(project);
             db.SaveChanges();
 
             TempData["message"] = "Proiectul a fost sters!";
             return RedirectToAction("Index");
-        }   
+        }
     }
 }
